Update index.html
We at Cocus are currently building a platform for self-sovereign identity, which is based on the Ethereum standards ERC 725/735.
We have a wide customer reach across several industries for current Blockchain solutions, e.g. Travel, Telco, Logistic, Food, Production, Automotive and Media & want to support the Open Source community as well as enhance the standard in terms of features, stability, services, applications, etc.